auth and cart fixes

+++ BookstoreASPNetServer/Controllers/CartController.cs
@@ -11,21 +11,35 @@ namespace BookstoreASPNetServer.Controllers
     public class CartController : ControllerBase
     {
         private readonly ICartRepository _cartRepository;
-        public CartController(ICartRepository cartRepository)
+        private readonly IAccountRepository _accountRepository;
+        public CartController(ICartRepository cartRepository, IAccountRepository accountRepository)
         {
             _cartRepository = cartRepository;
+            _accountRepository = accountRepository;
         }
 
         [HttpGet("{userId}")]
+        [Authorize]
         public async Task<IActionResult> GetCart([FromRoute] string userId)
         {
+            bool isValid = await _accountRepository.ValidateUserId(userId, User.Identity?.Name);
+            if (!isValid)
+            {
+                return Unauthorized();
+            }
             var cart = await _cartRepository.GetCartByUserId(userId);
             cart ??= new List<ProductInCartModel>();
             return Ok(cart);
         }
         [HttpPost("{userId}/add")]
+        [Authorize]
         public async Task<IActionResult> AddToCart([FromRoute] string userId, [FromBody] NewCartItemModel newCartItem)
         {
+            bool isValid = await _accountRepository.ValidateUserId(userId, User.Identity?.Name);
+            if (!isValid)
+            {
+                return Unauthorized();
+            }
             var result = await _cartRepository.AddProductToCart(userId, newCartItem);
             if (result == null)
             {
@@ -34,8 +48,14 @@ namespace BookstoreASPNetServer.Controllers
             return Ok(result);
         }
         [HttpPost("{userId}/addMany")]
+        [Authorize]
         public async Task<IActionResult> AddManyToCart([FromRoute] string userId, [FromBody] CartItemList cartItemList)
         {
+            bool isValid = await _accountRepository.ValidateUserId(userId, User.Identity?.Name);
+            if (!isValid)
+            {
+                return Unauthorized();
+            }
             if (cartItemList.CartItems.Count == 0)
             {
                 return BadRequest();
@@ -48,8 +68,14 @@ namespace BookstoreASPNetServer.Controllers
             return Ok(result);
         }
         [HttpDelete("{userId}/delete/{productId}")]
+        [Authorize]
         public async Task<IActionResult> RemoveProduct([FromRoute] string userId, [FromRoute] int productId)
         {
+            bool isValid = await _accountRepository.ValidateUserId(userId, User.Identity?.Name);
+            if (!isValid)
+            {
+                return Unauthorized();
+            }
             var result = await _cartRepository.RemoveProductFromCart(userId, productId);
             if (result == null)
             {
@@ -61,6 +87,11 @@ namespace BookstoreASPNetServer.Controllers
         [Authorize]
         public async Task<IActionResult> PlaceOrder([FromRoute] string userId)
         {
+            bool isValid = await _accountRepository.ValidateUserId(userId, User.Identity?.Name);
+            if (!isValid)
+            {
+                return Unauthorized();
+            }
             var result = await _cartRepository.PlaceOrder(userId);
             if (result == null)
             {
